(svn r24853) -Fix: Extend widget data member to 32 bits so that sprite IDs >= 2^16 can be used.

+++ src/newgrf_gui.cpp
@@ -1212,7 +1212,7 @@ struct NewGRFWindow : public Window, NewGRFScanCallback {
 			has_missing    |= c->status == GCS_NOT_FOUND;
 			has_compatible |= HasBit(c->flags, GCF_COMPATIBLE);
 		}
-		uint16 widget_data;
+		uint32 widget_data;
 		StringID tool_tip;
 		if (has_missing || has_compatible) {
 			widget_data = STR_NEWGRF_SETTINGS_FIND_MISSING_CONTENT_BUTTON;
